[untangleworms]: move read_params to modules
